--- src/Random.js
@@ -1,11 +1,0 @@
-const Random = {
-  getnerateRandomNumbers() {
-    const randomNumbers = [];
-    for (let i = 0; i < 6; i++) {
-      randomNumbers.push(Math.floor(Math.random() * 45) + 1);
-    }
-    return randomNumbers;
-  }
-};
-
-export default Random
